fix Accept header in HTTP requests
The ClientOperation.ProducesMediaTypes field determines the value of
the HTTP Accept header in requests.  The current implementation
causes /cacerts, /simpleenroll and /simplereenroll requests to be
sent with header `Accept: text/plain`.  This is incorrect, and can
result in EST servers reponding with status 406 Not Acceptable.  Use
the correct value: `application/pkcs7-mime`.

Signed-off-by: Fraser Tweedale <ftweedal@redhat.com>